refactor: extract shell action evaluation
